Move DummyCam instrument class to shared location to allow reuse
